move feed() to cis_env.py
