+++ cis_env.py
@@ -3,6 +3,7 @@ import random
 import numpy as np
 import dna_decoding
 from cis_cell import random_vector_of_length
+import scipy.constants as cons
 
 
 def curl(pos, vec_len=0.004, coor_origin=conf.CURL_CENTRE):
@@ -14,18 +15,56 @@ def curl(pos, vec_len=0.004, coor_origin=conf.CURL_CENTRE):
     return curl_pos * vec_len
 
 
-def feed_cell(cell, cell_dict, already_averaged_dict, food_factor=1):
-    f = random.uniform(0, 1)
-    if f < dna_decoding.food_theshold(cell.dna):
-        cell.energy_level += int(dna_decoding.food_amount(cell.dna)
-                                 * food_factor)
-    if conf.GENERAL_ENERGY_CONSUMPTION < cell.energy_level:
-        cell.energy_level -= conf.GENERAL_ENERGY_CONSUMPTION
-    else:
-        cell.energy_level = 0
+def feed(cell, time_step):
+    """
+        Give the cell food,
+        depending on position and time_step.
+    """
 
-    average_out_energy_in_connected_cells(
-        cell, cell_dict, already_averaged_dict)
+    #  get value from the food_function --> definition area: {-3, 3}
+    food_value = food_function(cell.pos, time_step)
+
+    # normalize food_value to {0, 1}
+    food_value = normalize(food_value)
+
+    # check if cell gets (the fixed amount of) food
+    if food_value < conf.FOOD_THRESHOLD:
+        cell.energy_level += conf.FOOD_ENERGY
+
+    return cell
+
+
+def normalize(num, definition_area_size=6):
+    """
+        Normalizes the defenition area of the num to {0, 1} .
+    """
+    ret = round(num) / definition_area_size + 0.5
+    return ret
+
+
+def food_function(cell_pos, time_step):
+    """
+        Creates the 4-dim food function and returns it.
+    """
+    x = wave_function(cell_pos.x, time_step)
+    y = wave_function(cell_pos.y, time_step)
+    z = wave_function(cell_pos.z, time_step)
+
+    return x + y + z
+
+
+def wave_function(x, t, max_ampli=1, oscillation_period=1, init_deflection=0):
+    """
+        Classical mechanical wave function.
+    """
+
+    time_relation = t / oscillation_period
+    space_relation = x / (cons.speed_of_light * oscillation_period)
+
+    # sinput = sin + input
+    sinput = 2 * np.pi * (time_relation - space_relation) + init_deflection
+
+    return max_ampli * np.sin(sinput)
 
 
 def get_value(dict, key):
